+++ internal/renderers/femtovg/Cargo.toml
@@ -1,0 +1,63 @@
+# Copyright © SixtyFPS GmbH <info@slint-ui.com>
+# SPDX-License-Identifier: GPL-3.0-only OR LicenseRef-Slint-commercial
+
+[package]
+name = "i-slint-renderer-femtovg"
+version = "1.0.0"
+authors = ["Slint Developers <info@slint-ui.com>"]
+edition = "2021"
+license = "GPL-3.0-only OR LicenseRef-Slint-commercial"
+description = "FemtoVG based renderer for Slint"
+repository = "https://github.com/slint-ui/slint"
+homepage = "https://slint-ui.com"
+rust-version.workspace = true
+
+[lib]
+path = "lib.rs"
+
+[features]
+diskfonts = ["fontdb/memmap", "fontdb/fontconfig"]
+fontconfig = ["yeslogic-fontconfig-sys"]
+
+[dependencies]
+i-slint-core = { version = "=1.0.0", path = "../../../internal/core", features = [ "box-shadow-cache" ] }
+i-slint-core-macros = { version = "=1.0.0", path = "../../../internal/core-macros" }
+i-slint-common = { version = "=1.0.0", path = "../../../internal/common" }
+
+const-field-offset = { version = "0.1", path = "../../../helper_crates/const-field-offset" }
+vtable = { version = "0.1.6", path = "../../../helper_crates/vtable" }
+
+cfg-if = "1"
+derive_more = "0.99.5"
+lyon_path = "1.0"
+once_cell = "1.5"
+pin-weak = "1"
+scoped-tls-hkt = "0.1"
+instant = "0.1"
+raw-window-handle = { version = "0.5", features = ["alloc"] }
+femtovg = { version = "0.6.0" }
+fontdb = { version = "0.12.0", default-features = false }
+ttf-parser = { version = "0.18.0" } # Use the same version was femtovg's rustybuzz, to avoid duplicate crates
+unicode-script = { version = "0.5.4" } # Use the same version was femtovg's rustybuzz, to avoid duplicate crates
+imgref = { version = "1.6.1" }
+rgb = { version = "0.8.27" }
+
+glow = { version = "0.12" }
+unicode-segmentation = { version = "1.8.0" }
+
+[target.'cfg(target_family = "windows")'.dependencies]
+winapi = { version = "0.3", features = ["dwrite"] }
+dwrote = { version = "0.11.0" }
+
+[target.'cfg(target_os = "macos")'.dependencies]
+# For GL rendering
+core-foundation = { version = "0.9.1" }
+core-text = { version = "19.1.0" }
+
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+web-sys = { version = "0.3", features=["console", "WebGlContextAttributes", "CanvasRenderingContext2d", "HtmlInputElement", "HtmlCanvasElement", "Window", "Document"] }
+wasm-bindgen = { version = "0.2" }
+
+[target.'cfg(not(any(target_family = "windows", target_os = "macos", target_os = "ios", target_arch = "wasm32")))'.dependencies]
+libc = { version = "0.2" }
+yeslogic-fontconfig-sys = { version = "3.2", optional = true }
