+++ internal/renderers/femtovg/README.md
@@ -1,0 +1,6 @@
+**NOTE**: This library is an **internal** crate of the [Slint project](https://slint-ui.com).
+This crate should **not be used directly** by applications using Slint.
+You should use the `slint` crate instead.
+
+**WARNING**: This crate does not follow the semver convention for versioning and can
+only be used with `version = "=x.y.z"` in Cargo.toml.
