+++ scripts/publish.sh
@@ -10,6 +10,7 @@ cargo publish --manifest-path internal/core/Cargo.toml
 sleep 15
 cargo publish --manifest-path api/rs/macros/Cargo.toml
 cargo publish --manifest-path internal/renderers/skia/Cargo.toml
+cargo publish --manifest-path internal/renderers/femtovg/Cargo.toml
 cargo publish --manifest-path internal/backends/winit/Cargo.toml --features x11,renderer-winit-femtovg
 cargo publish --manifest-path api/rs/build/Cargo.toml
 cargo publish --manifest-path internal/backends/qt/Cargo.toml
